requirements.txt
django
psycopg2-binary
djangorestframework
django-cors-headers
minio